requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
greenlet==3.3.2
idna==3.11
iniconfig==2.3.0
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
